// code-app/src/generated/models/Nvm_assetsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Nvm_assetsnvm_criticality = {
  877010000: 'Low',
  877010001: 'Medium',
  877010002: 'High',
  877010003: 'Critical'
} as const;
export type Nvm_assetsnvm_criticality = keyof typeof Nvm_assetsnvm_criticality;
export const Nvm_assetsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Nvm_assetsstatecode = keyof typeof Nvm_assetsstatecode;
export const Nvm_assetsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Nvm_assetsstatuscode = keyof typeof Nvm_assetsstatuscode;

export interface Nvm_assetsBase {
  importsequencenumber?: number;
  nvm_assetid: string;
  nvm_assetname?: string;
  nvm_assettag?: string;
  nvm_criticality?: Nvm_assetsnvm_criticality;
  nvm_installdate?: string;
  nvm_manufacturer?: string;
  nvm_planx?: number;
  nvm_plany?: number;
  nvm_replacementcost?: number;
  "nvm_Site@odata.bind"?: string;
  overriddencreatedon?: string;
  statecode: Nvm_assetsstatecode;
  statuscode?: Nvm_assetsstatuscode;
  timezoneruleversionnumber?: number;
  "transactioncurrencyid@odata.bind"?: string;
  utcconversiontimezonecode?: number;
}

export interface Nvm_assets extends Nvm_assetsBase {
  createdon?: string;
  exchangerate?: number;
  modifiedon?: string;
  nvm_criticalityname?: string;
  nvm_replacementcost_base?: number;
  ownerid: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  nvm_site?: object;
  _nvm_site_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}